backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
# psycopg2-binary==2.9.9  # Only needed for PostgreSQL in production

# Testing
pytest==7.4.4
pytest-django==4.7.0
coverage==7.4.0

# Code quality
flake8==7.0.0
black==24.1.1

# Production server
gunicorn==21.2.0
whitenoise==6.6.0